CMFIVE-53: Реализация наследование типов доменных объектов

+++ CM5/Core/dao-api/src/main/java/ru/intertrust/cm/core/dao/api/DomainObjectTypeIdDao.java
@@ -27,5 +27,5 @@ public interface DomainObjectTypeIdDao {
      * @param domainObjectTypeName
      * @return
      */
-    public Long insert(String domainObjectTypeName);
+    public Integer insert(String domainObjectTypeName);
 }
